Ensure attn bias is the same dtype as query
Accidentally removed this change in commit aed2e5deb7f50fb1c0b8bb0a399d44c08e652799

+++ fabric/weighted_attn.py
@@ -110,5 +110,5 @@ def get_attn_bias(weights, shape, dtype):
     min_val = torch.finfo(dtype).min
     attn_bias = weights.log().clamp(min=min_val)
     view_shape = [B, *([1] * (len(shape) - 2)), nk]
-    attn_bias = attn_bias.view(view_shape).expand(shape)
+    attn_bias = attn_bias.view(view_shape).expand(shape).to(dtype)
     return attn_bias
